Allow Travis to fail on Julia Nightly build
Nightly builds are unstable. It would make more sense to be aware that the nightly build failed,
without worrying about it too much, than try to get our PRs to pass the nightly build.
I changed the Travis config file to simply allow for the PR checks to pass even if the nightly build failed.